requirements.txt
Django==2.2.1
django-cors-headers==3.0.2
django-filter==2.1.0
djangorestframework==3.9.4
drf-flex-fields==0.5.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
python-dotenv==0.10.2
pytz==2019.1
sqlparse==0.3.0
